Simulation moved to a special class
